PPLUS: Integrator as part of the flow
