+++ composer.json
@@ -13,6 +13,7 @@
         "knplabs/github-api": "^3.4",
         "nikic/php-parser": "^4.13",
         "phpstan/phpstan": "^1.2",
+        "spryker-sdk/integrator": "1.0.x-dev",
         "spryker-sdk/sdk-contracts": "dev-master",
         "spryker/code-sniffer": "^0.17.3",
         "symfony/config": "^5.0",
@@ -51,7 +52,8 @@
     "config": {
         "sort-packages": true,
         "allow-plugins": {
-            "dealerdirect/phpcodesniffer-composer-installer": true
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "spryker-sdk/integrator": true
         }
     },
     "scripts": {
