--- src/Evaluate/Infrastructure/DependencyInjection/config/services.yml
@@ -1,25 +1,0 @@
-services:
-    _defaults:
-        autowire: true
-        autoconfigure: true
-        public: true
-
-# Evaluate
-    Evaluate\:
-        resource: '../../../../Evaluate/*'
-
-    analyze.task:
-        class: Evaluate\Infrastructure\Task\Analyze\AnalyzeTask
-        tags: [ 'sdk.task' ]
-
-    report.task:
-        class: Evaluate\Infrastructure\Task\Report\ReportTask
-        tags: [ 'sdk.task' ]
-
-    report.listener:
-        class: Evaluate\Infrastructure\Event\ReportListener
-        tags:
-              - { name: kernel.event_listener, event: console.command, method: onConsoleCommand }
-
-# Third-party
-    Symfony\Component\Yaml\Yaml: ~
