--- src/Evaluate/Infrastructure/Lifecycle/Lifecycle.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace Evaluate\Infrastructure\Lifecycle;
-
-use SprykerSdk\SdkContracts\Entity\Lifecycle\LifecycleInterface;
-
-class Lifecycle implements LifecycleInterface
-{
-}
