+++ src/Upgrader/Commands/Integrator/IntegratorCommand.php
@@ -1,0 +1,63 @@
+<?php
+
+/**
+ * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+namespace Upgrader\Commands\Integrator;
+
+use SprykerSdk\SdkContracts\Entity\CommandInterface;
+use SprykerSdk\SdkContracts\Entity\ContextInterface;
+use SprykerSdk\SdkContracts\Entity\ConverterInterface;
+
+class IntegratorCommand implements CommandInterface
+{
+    /**
+     * @return string
+     */
+    public function getCommand(): string
+    {
+        return `vendor/bin/integrator`;
+    }
+
+    /**
+     * @return string
+     */
+    public function getType(): string
+    {
+        return 'php';
+    }
+
+    /**
+     * @return array<string>
+     */
+    public function getTags(): array
+    {
+        return [];
+    }
+
+    /**
+     * @return bool
+     */
+    public function hasStopOnError(): bool
+    {
+        return true;
+    }
+
+    /**
+     * @return \SprykerSdk\SdkContracts\Entity\ConverterInterface|null
+     */
+    public function getViolationConverter(): ?ConverterInterface
+    {
+        return null;
+    }
+
+    /**
+     * @return string
+     */
+    public function getStage(): string
+    {
+        return ContextInterface::DEFAULT_STAGE;
+    }
+}
