+++ src/Upgrader/Lifecycle/Lifecycle.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+namespace Upgrader\Lifecycle;
+
+use SprykerSdk\SdkContracts\Entity\Lifecycle\LifecycleInterface;
+
+class Lifecycle implements LifecycleInterface
+{
+}
